changes on all algorithm dialogs UI components
