+++ LDMP/calculate.py
@@ -21,7 +21,6 @@ from osgeo import (
     gdal,
     ogr,
 )
-
 from PyQt5 import (
     QtCore,
     QtGui,
@@ -835,6 +834,59 @@ class DlgCalculateBase(QtWidgets.QDialog):
     @classmethod
     def get_subclass_name(cls):
         return cls.__name__
+
+    def splitter_toggled(self):
+        if self.splitter_collapsed:
+            self.splitter.restoreState(self.splitter_state)
+            self.collapse_button.setArrowType(Qt.RightArrow)
+        else:
+            self.splitter_state = self.splitter.saveState()
+            self.splitter.setSizes([1, 0])
+            self.collapse_button.setArrowType(Qt.LeftArrow)
+        self.splitter_collapsed = not self.splitter_collapsed
+
+    def initiliaze_settings(self):
+
+        ok_button = self.button_box.button(
+            QtWidgets.QDialogButtonBox.Ok
+        )
+        ok_button.setText(self.tr("Schedule execution"))
+        ok_button.clicked.connect(self.btn_calculate)
+
+        region = QgsSettings().value("trends_earth/region_of_interest/area_settings_name")
+        self.execution_name_le.setText(f"LDN_{region}_{datetime.strftime(datetime.now(), '%Y%m%d%H%M%S')}")
+
+        self.region_button.clicked.connect(self.run_settings)
+
+        # adding a collapsible arrow on the splitter
+        self.splitter_state = self.splitter.saveState()
+        self.splitter.setCollapsible(0, False)
+        splitter_handle = self.splitter.handle(1)
+        handle_layout = QtWidgets.QVBoxLayout()
+        handle_layout.setContentsMargins(0, 0, 0, 0)
+        self.collapse_button = QtWidgets.QToolButton(splitter_handle)
+        self.collapse_button.setAutoRaise(True)
+        self.collapse_button.setFixedSize(12, 12)
+        self.collapse_button.setCursor(Qt.ArrowCursor)
+        handle_layout.addWidget(self.collapse_button)
+
+        handle_layout.addStretch()
+        splitter_handle.setLayout(handle_layout)
+
+        arrow_type = Qt.RightArrow if self.splitter.sizes()[1] == 0 else Qt.LeftArrow
+
+        self.collapse_button.setArrowType(arrow_type)
+        self.collapse_button.clicked.connect(self.splitter_toggled)
+        self.splitter_collapsed = self.splitter.sizes()[1] != 0
+
+        QgsGui.enableAutoGeometryRestore(self)
+
+        self.region_la.setText(self.tr(f"Current region: {region}"))
+
+    def run_settings(self):
+        dlg_settings = DlgSettings()
+        dlg_settings.show()
+        result = dlg_settings.exec_()
 
     def showEvent(self, event):
         super(DlgCalculateBase, self).showEvent(event)
