+++ LDMP/gui/DlgCalculateOneStep.ui
@@ -304,8 +304,8 @@
               <rect>
                <x>0</x>
                <y>0</y>
-               <width>772</width>
-               <height>16</height>
+               <width>786</width>
+               <height>187</height>
               </rect>
              </property>
             </widget>
